aosp: use the Android system fonts
Stop resolving the font directories from the Evergreen content path, added in
https://github.com/youtube/cobalt/pull/11336, and implement the Evergreen font
layout instead: the platform provides the fonts at
kSbSystemPathContentDirectory/fonts/, while the system image content directory
at kSbSystemPathContentDirectory/app/cobalt/content/fonts/ contains the "empty"
font package.

Modify the system font directory to point to the platform fonts and set
cobalt_font_package to "android_system".

Lastly, add font fallback path resolution for access() to the Android posix
emulation layer, so that fonts are searched at /system/fonts when they aren't
found at the system font directory.

Bug: 495203133
