+++ .gn
@@ -175,7 +175,7 @@ exec_script_allowlist += [
   "//starboard/tools/BUILD.gn",
   "//starboard/build/config/sabi/BUILD.gn",
   "//starboard/build/platform_path.gni",
-  "//starboard/content/fonts/BUILD.gn",
+  "//starboard/content/fonts/variables.gni",
 ]
 
 # TODO(cobalt, b/377295011): figure out a better long-term solution here, or
