+++ starboard/android/shared/system_get_path.cc
@@ -78,7 +78,7 @@ bool SbSystemGetPath(SbSystemPathId path_id, char* out_path, int path_size) {
     case kSbSystemPathFontConfigurationDirectory:
     case kSbSystemPathFontDirectory:
 #if BUILDFLAG(IS_STARBOARD)
-      if (starboard::strlcpy(path, GetContentPath(), kPathSize) >= kPathSize) {
+      if (starboard::strlcpy(path, g_app_assets_dir, kPathSize) >= kPathSize) {
         return false;
       }
       if (starboard::strlcat(path, "/fonts", kPathSize) >= kPathSize) {
