+++ starboard/nplb/nplb_evergreen_compat_tests/fonts_test.cc
@@ -26,10 +26,6 @@
 #error These tests apply only to EVERGREEN_COMPATIBLE platforms.
 #endif
 
-#if defined(ANDROID)
-#error These tests are not applicable to AOSP
-#endif
-
 namespace nplb {
 
 namespace {
